Move models into `web-common/features/models` directory (#1554)
* Move models into `web-common/features/models` directory

* Move `query-highlight-store` to `feature/models`

* Move model-specific utils to `feature/models`
