+++ web-common/src/features/models/index.ts
@@ -1,0 +1,2 @@
+export { default as ModelAssets } from "./navigation/ModelAssets.svelte";
+export { default as ModelWorkspace } from "./workspace/ModelWorkspace.svelte";
