+++ web-common/src/features/models/navigation/ModelTooltip.svelte
