+++ web-common/src/features/models/query-highlight-store.ts
