+++ web-common/src/features/models/utils/get-table-references/get-table-references.test.ts
@@ -1,4 +1,4 @@
-import { getTableReferences } from "./";
+import { getTableReferences } from ".";
 import { tests } from "./test-data";
 
 describe("getAllTableReferences", () => {
