+++ web-common/src/features/models/workspace/inspector/WithModelResultTooltip.svelte
