+++ web-common/src/features/sources/createModel.ts
@@ -1,4 +1,4 @@
-import { createModel } from "@rilldata/web-local/lib/components/navigation/models/createModel";
+import { createModel } from "@rilldata/web-common/features/models/createModel";
 import { getName } from "@rilldata/web-local/lib/util/incrementName";
 import type { QueryClient, UseMutationResult } from "@sveltestack/svelte-query";
 import { notifications } from "../../components/notifications";
