+++ web-local/src/lib/components/navigation/Navigation.svelte
@@ -5,6 +5,7 @@
   import Portal from "@rilldata/web-common/components/Portal.svelte";
   import Tooltip from "@rilldata/web-common/components/tooltip/Tooltip.svelte";
   import TooltipContent from "@rilldata/web-common/components/tooltip/TooltipContent.svelte";
+  import { ModelAssets } from "@rilldata/web-common/features/models";
   import TableAssets from "@rilldata/web-common/features/sources/navigation/TableAssets.svelte";
   import { useRuntimeServiceGetFile } from "@rilldata/web-common/runtime-client";
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
@@ -17,7 +18,6 @@
   import { drag } from "../../drag";
   import MetricsDefinitionAssets from "./dashboards/MetricsDefinitionAssets.svelte";
   import Footer from "./Footer.svelte";
-  import ModelAssets from "./models/ModelAssets.svelte";
   import { shorthandTitle } from "./shorthand-title";
 
   let mounted = false;
