+++ web-local/src/lib/components/workspace/index.ts
@@ -6,4 +6,3 @@ export { default as WorkspaceHeaderStatusSpinner } from "./core/WorkspaceHeaderS
 /** individual route elements */
 export { default as DashboardWorkspace } from "./explore/Explore.svelte";
 export { default as MetricsDefinitionWorkspace } from "./metrics-def/MetricsDefWorkspace.svelte";
-export { default as ModelWorkspace } from "./model/ModelWorkspace.svelte";
