+++ web-local/src/lib/components/workspace/metrics-def/MetricsDefModelSelector.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
+  import { useModelNames } from "@rilldata/web-common/features/models/selectors";
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
-  import { useModelNames } from "@rilldata/web-local/lib/svelte-query/models";
   import type { Readable } from "svelte/store";
   import type { MetricsInternalRepresentation } from "../../../application-state-stores/metrics-internal-store";
